Update for Gnome Shell 3.36.
The system menu implementation has completely changed in GS 3.36.
The settings item is now a menu entry and the icons are gone.
Update the extension to deal with a menu entry instead of a button.
It's not possible anymore to merge the Settings and Tweaks entries.
Remove the "merge-with-settings" preference.